Simplifications and widget grid removal
WidgetGrid component has been removed from claroline :
As in the home plugin of claroline,  it is now replaced by the rendering
of each widget present in the corresponding array.

References to modules files within the module now follow the convention
provided in claroline webpack "vendor-shortcut" plugin.

Some code cleaning applied.

Context of rendering is now set to be the current workspace.